Extract deploy stage plan support
